Code refactor with reducer and new action type SET_ERROR.
