+++ client/marketing/data/action-types.js
@@ -1,0 +1,5 @@
+const TYPES = {
+	SET_ERROR: 'SET_ERROR',
+};
+
+export default TYPES;
